+++ R/SVG.R
@@ -4,11 +4,12 @@
 #' The driver output is currently NOT specifying a DOCTYPE DTD.
 #'
 #' @param file The file where output will appear. There are two special values
-#'   that can be recognized by the device: \code{file = ":terminal:"}
-#'   prints the content of SVG to the R terminal, and \code{file = ":string:"}
-#'   saves the SVG to a character string which can be retrieved by
-#'   \code{\link{stringSVG}()} after closing the device. These two special values
-#'   will not generate temporary files on the hard disk.
+#'   recognized by the device:
+#'
+#'   \code{file = ":terminal:"} prints the content of SVG to the R terminal.
+#'
+#'   \code{file = ":string:"} saves the SVG to a character string which can be
+#'   retrieved by \code{\link{stringSVG}()} after closing the device.
 #' @param height,width Height and width in inches.
 #' @param bg Default background color for the plot (defaults to "white").
 #' @param pointsize Default point size.
@@ -21,22 +22,21 @@
 #'   \email{matthieu.decorde@@ens-lyon.fr}
 #' @seealso \code{\link{pictex}}, \code{\link{postscript}}, \code{\link{Devices}}
 #' @examples
-#' ## Save to file
+#' # Save to file
 #' svglite("Rplots.svg")
 #' plot(1:11, (-5:5)^2, type = 'b', main = "Simple Example")
 #' dev.off()
 #'
-#' ## Print to R console
+#' # Print to R console
 #' svglite(":terminal:")
 #' plot.new()
 #' dev.off()
 #'
-#' ## Write to a string stream
+#' # Write to a string stream
 #' svglite(":string:")
 #' plot.new()
 #' dev.off()
-#' svgstr <- stringSVG()
-#' cat(svgstr)
+#' cat(stringSVG())
 #'
 #' @keywords device
 #' @useDynLib svglite
@@ -49,26 +49,24 @@ svglite <- function(file = "Rplots.svg", width = 10, height = 8, bg = "white",
   invisible(devSVG_(file, bg, width, height, pointsize, standalone))
 }
 
-
 .pkg_env <- new.env()
 ## This string will be modified by device functions
 .pkg_env$svg_string <- ""
 
-
-#' Extract the content of SVG as a string
+#' Access current SVG as a string.
 #'
 #' This function is used to extract the content of SVG produced by
 #' the most recent call of \code{svglite(":string:")},
 #' which directly writes the SVG to a string stream
 #' and does not generate any temporary files.
 #'
-#' @return A character string representing the last plot created by \code{svglite(":string:")}.
+#' @return A character string representing the last plot created by
+#'   \code{svglite(":string:")}.
 #' @examples
 #' svglite(":string:")
 #' plot(rnorm(5), rnorm(5))
 #' dev.off()
-#' svgstr = stringSVG()
-#' cat(svgstr)
+#' cat(stringSVG())
 #' @export
 stringSVG <- function() {
   .pkg_env$svg_string